backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
fastapi_camelcase==2.0.0
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
gotrue==2.12.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
openai==1.97.1
orjson==3.11.0
packaging==25.0
postgrest==1.1.1
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
Pygments==2.19.2
pyhumps==3.8.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
realtime==2.6.0
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.8
rignore==0.6.4
sentry-sdk==2.33.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
storage3==0.12.0
StrEnum==0.4.15
supabase==2.17.0
supafunc==0.10.1
tenacity==9.1.2
termcolor==3.1.0
tiktoken==0.9.0
tqdm==4.67.1
typer==0.16.0
typing_extensions==4.14.1
typing-inspection==0.4.1
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
